move Activity to folder
